+++ test/index_test.js
@@ -110,6 +110,59 @@ describe("Objectware", function() {
     })
   })
 
+  describe(".cloneDeep", function() {
+    it("must return undefined given nothing", function() {
+      demand(_.cloneDeep()).be.undefined()
+    })
+
+    it("must return undefined given undefined", function() {
+      demand(_.cloneDeep(undefined)).be.undefined()
+    })
+
+    it("must return null given null", function() {
+      demand(_.cloneDeep(null)).be.null()
+    })
+
+    it("must return new object", function() {
+      var obj = {}
+      _.cloneDeep(obj).must.not.equal(obj)
+    })
+
+    it("must return new object given nested objects", function() {
+      var obj = {attributes: {}}
+      _.cloneDeep(obj).attributes.must.not.equal(obj.attributes)
+    })
+
+    it("must clone properties", function() {
+      _.cloneDeep({name: "John", age: 42}).must.eql({name: "John", age: 42})
+    })
+
+    it("must clone nested properties", function() {
+      var obj = {name: "John", attributes: {age: 13}}
+      _.cloneDeep(obj).must.eql({name: "John", attributes: {age: 13}})
+    })
+
+    it("must assign functions", function() {
+      function fn() {}
+      _.cloneDeep({fn: fn}).must.eql({fn: fn})
+    })
+
+    it("must clone properties from inherited sources", function() {
+      _.cloneDeep(Object.create({name: "John"})).must.eql({name: "John"})
+    })
+
+    it("must not clone unenumerable properties", function() {
+      var source = Object.defineProperty({}, "name", {value: "John"})
+      _.cloneDeep(source).must.eql({})
+    })
+
+    // Just to ensure the target isn't shared between invocations.
+    it("must clone properties when called twice", function() {
+      _.cloneDeep({name: "John"})
+      _.cloneDeep({age: 42}).must.eql({age: 42})
+    })
+  })
+
   describe(".filter", function() {
     function isEven(value) { return value % 2 == 0 }