+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - Fixes [`Objectware.merge`][merge] to clone plain objects objects
   recursively.  
   Previously in some cases it cloned objects shallowly.
+- Adds [`Objectware.cloneDeep`][cloneDeep].
+
+[cloneDeep]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.cloneDeep
 
 ## 1.1.0 (Feb 7, 2015)
 - Adds [`Objectware.isPlainObject`][isPlainObject].